+++ NEWS.md
@@ -1,5 +1,6 @@
 # tidychangepoint 1.0.1
 
+* Added support for model from the `segmented` package.
 * `fitness.wbs()` returns the MBIC penalty value.
 * Added `summary()` methods for `tidycpt`, `seg_cpt`, and `mod_cpt` objects.
 * Improved output from `print()` methods.