+++ NEWS.md
@@ -1,6 +1,8 @@
 # tidychangepoint 1.0.1
 
-* Added support for model from the `segmented` package.
+* Added preliminary support for ARIMA models. 
+* Added support for models from the `changepointGA` package.
+* Added support for models from the `segmented` package.
 * `fitness.wbs()` returns the MBIC penalty value.
 * Added `summary()` methods for `tidycpt`, `seg_cpt`, and `mod_cpt` objects.
 * Improved output from `print()` methods.